fix: DirectoryBrowserSupportTest
